refactor(typ): migrate provider types to public packages
Migrate Provider, AuthType, and OAuthDetail types from internal/typ to
public packages following the new migration pattern:

- Create protocol/ package with aliases to internal/protocol types
- Create common/provider/ package for provider-related types
- Update internal/typ/type.go to use aliases for backward compatibility

This establishes a pattern for future migrations of internal types to
public API while maintaining zero breaking changes through type aliases.

Related: #802
